refactor: Consolidate onboarding screens and fix build errors
- Fixed missing Icons import in MacroStatsRow.kt
- Fixed missing @Preview import in GenderScreen.kt
- Fixed wrong import paths in app DashboardViewModel.kt
- Updated MainActivity imports to use feature modules
- Created UserProfileScreen.kt with 6-step horizontal pager
- Updated OnboardingNavHost to use consolidated UserProfileScreen
- Added selectedDate and water tracking to feature DashboardViewModel
- Reduced onboarding navigation from 20+ to 15 destinations

+++ feature-dashboard/src/main/java/com/example/calview/feature/dashboard/components/CalorieRing.kt
@@ -16,8 +16,8 @@ import com.example.calview.core.ui.theme.CalAIGreen
 
 @Composable
 fun CalorieRing(
-    consumed: Int,
-    goal: Int,
+    consumed: Float,
+    goal: Float,
     modifier: Modifier = Modifier
 ) {
     val progress = (consumed.toFloat() / goal.toFloat()).coerceIn(0f, 1f)
